pyproject.toml
[tool.poetry]
description = "Terraria Discord AFK Points Farm"
authors = ["FoxFil <filipp.vesker@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
